+++ osidb/urls.py
@@ -6,7 +6,7 @@ from django.urls import include, path, re_path
 from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
 from rest_framework import routers
 
-from apps.workflows.api import promote, reject
+from apps.workflows.api import ResetWorkflow, RevertWorkflow, promote, reject
 from osidb.helpers import get_execution_env
 
 from .api_views import (
@@ -96,6 +96,14 @@ urlpatterns = [
     re_path(
         rf"^api/{OSIDB_API_VERSION}/flaws/(?P<flaw_id>[^/.]+)/promote$",
         promote.as_view(),
+    ),
+    re_path(
+        rf"^api/{OSIDB_API_VERSION}/flaws/(?P<flaw_id>[^/.]+)/revert$",
+        RevertWorkflow.as_view(),
+    ),
+    re_path(
+        rf"^api/{OSIDB_API_VERSION}/flaws/(?P<flaw_id>[^/.]+)/reset$",
+        ResetWorkflow.as_view(),
     ),
     re_path(
         rf"^api/{OSIDB_API_VERSION}/flaws/(?P<flaw_id>[^/.]+)/reject$",